The dutch translation is only partial visible on the full form.
To fix this I added 1 extra row to the warningLabel so more text can be displayed.

and the sendAndQuitButton is now 50% wider so all translations are visible within this button. Most translations need more horizontal space then the default English translation 'Send and quit'.
The text within the button is centered instead of right aligned.