Update Mail.cs
Fix bug where the email port is not read when the default port and use ssl checkboxes are checked.

When ssl was checked the port would be 25 even thought in the UI the correct value was visible.

The solution is the always read from the portNumericUpDown. Since the portNumericUpDown control always contains a value.