* Fixing Issue #68 NBug.Helpers.EmailDestinationBuilder is creating a connection string with yes and no instead of true and false.
Fields are case sensitive Replace useSSL with UseSsl. UseSsl is a property of Mail.cs

The EmailDestinationBuilder is not used in the project? So the real fix is updating the example connection string in mail.cs

Adding xunit.runner.visualstudio to the test project so xunit tests can be run in visual studio 2015.